+++ pytorch_segmentation_models_trainer/dataset_loader/dataset.py
@@ -18,7 +18,9 @@
  *                                                                         *
  ****
 """
+from abc import abstractmethod
 import os
+from pytorch_segmentation_models_trainer.utils.object_detection_utils import bbox_xywh_to_xyxy, bbox_xyxy_to_xywh
 from albumentations.pytorch import ToTensorV2
 from albumentations.augmentations.transforms import Normalize
 import torch
@@ -33,22 +35,20 @@ from PIL import Image
 from torch.utils.data import Dataset
 import json
 
+from omegaconf import DictConfig, OmegaConf
 
-def load_augmentation_object(input_list):
+
+def load_augmentation_object(input_list, bbox_params=None):
     try:
         aug_list = [
-            instantiate(i) for i in input_list
+            instantiate(i, _recursive_=False) for i in input_list
         ]
     except:
         aug_list = input_list
-    return A.Compose(aug_list)
+    return A.Compose(aug_list) if bbox_params is None \
+        else A.Compose(aug_list, bbox_params=OmegaConf.to_container(bbox_params))
 
-class SegmentationDataset(Dataset):
-    """[summary]
-
-    Args:
-        Dataset ([type]): [description]
-    """
+class AbstractDataset(Dataset):
     def __init__(
         self,
         input_csv_path: Path,
@@ -73,19 +73,17 @@ class SegmentationDataset(Dataset):
     def __len__(self) -> int:
         return self.len
 
+    @abstractmethod
     def __getitem__(self, idx: int) -> Dict[str, Any]:
-        idx = idx % self.len
+        """Item getter. Must be reimplemented in each subclass.
 
-        image = self.load_image(idx, key=self.image_key)
-        mask = self.load_image(idx, key=self.mask_key, is_mask=True)
-        result = {
-            'image': image,
-            'mask': mask
-        } if self.transform is None else self.transform(
-                image=image,
-                mask=mask
-            )
-        return result
+        Args:
+            idx (int): index of the item to be returned
+
+        Returns:
+            Dict[str, Any]: Loaded item.
+        """
+        pass
 
     def get_path(self, idx, key=None):
         key = self.image_key if key is None else key
@@ -107,6 +105,50 @@ class SegmentationDataset(Dataset):
             image = image.convert('RGB')
         image = np.array(image)
         return (image > 0).astype(np.uint8) if is_mask else image
+    
+    def to_tensor(self, x):
+        return x if isinstance(x, torch.Tensor) \
+            else torch.from_numpy(x)
+
+class SegmentationDataset(AbstractDataset):
+    """[summary]
+
+    Args:
+        Dataset ([type]): [description]
+    """
+    def __init__(
+        self,
+        input_csv_path: Path,
+        root_dir=None,
+        augmentation_list=None,
+        data_loader=None,
+        image_key=None,
+        mask_key=None,
+        n_first_rows_to_read=None
+    ) -> None:
+        super(SegmentationDataset, self).__init__(
+            input_csv_path=input_csv_path,
+            root_dir=root_dir,
+            augmentation_list=augmentation_list,
+            data_loader=data_loader,
+            image_key=image_key,
+            mask_key=mask_key,
+            n_first_rows_to_read=n_first_rows_to_read
+        )
+
+    def __getitem__(self, idx: int) -> Dict[str, Any]:
+        idx = idx % self.len
+
+        image = self.load_image(idx, key=self.image_key)
+        mask = self.load_image(idx, key=self.mask_key, is_mask=True)
+        result = {
+            'image': image,
+            'mask': mask
+        } if self.transform is None else self.transform(
+                image=image,
+                mask=mask
+            )
+        return result
 
 class FrameFieldSegmentationDataset(SegmentationDataset):
     def __init__(
@@ -134,9 +176,11 @@ class FrameFieldSegmentationDataset(SegmentationDataset):
         gpu_augmentation_list=None
     ) -> None:
         mask_key = 'polygon_mask' if mask_key is None else mask_key
-        super().__init__(input_csv_path, root_dir=root_dir, augmentation_list=augmentation_list,
-                         data_loader=data_loader, image_key=image_key, mask_key=mask_key,
-                         n_first_rows_to_read=n_first_rows_to_read)
+        super(FrameFieldSegmentationDataset, self).__init__(
+            input_csv_path, root_dir=root_dir, augmentation_list=augmentation_list,
+            data_loader=data_loader, image_key=image_key, mask_key=mask_key,
+            n_first_rows_to_read=n_first_rows_to_read
+        )
         self.multi_band_mask = multi_band_mask
         self.boundary_mask_key = boundary_mask_key if boundary_mask_key is not None else 'boundary_mask'
         self.vertex_mask_key = vertex_mask_key if vertex_mask_key is not None else 'vertex_mask'
@@ -184,9 +228,6 @@ class FrameFieldSegmentationDataset(SegmentationDataset):
     def compute_class_freq(self, gt_polygons_image):
         pass
 
-    def to_tensor(self, x):
-        return x if isinstance(x, torch.Tensor) \
-            else torch.from_numpy(x)
     
     def get_mean_axis(self, mask):
         if len(mask.shape) > 2:
@@ -279,7 +320,7 @@ class FrameFieldSegmentationDataset(SegmentationDataset):
         return ds_item_dict
         
 
-class PolygonRNNDataset(Dataset):
+class PolygonRNNDataset(AbstractDataset):
     def __init__(
         self,
         input_csv_path: Path,
@@ -291,26 +332,16 @@ class PolygonRNNDataset(Dataset):
         mask_key=None,
         n_first_rows_to_read=None
     ) -> None:
-        super().__init__()
-        self.input_csv_path = input_csv_path
-        self.root_dir = root_dir
+        super(PolygonRNNDataset, self).__init__(
+            input_csv_path=input_csv_path,
+            root_dir=root_dir,
+            augmentation_list=augmentation_list,
+            data_loader=data_loader,
+            image_key=image_key,
+            mask_key=mask_key,
+            n_first_rows_to_read=n_first_rows_to_read
+        )
         self.sequence_length = sequence_length
-        self.df = pd.read_csv(input_csv_path) if n_first_rows_to_read is None \
-            else pd.read_csv(input_csv_path, nrows=n_first_rows_to_read)
-        self.transform = None if augmentation_list is None \
-            else load_augmentation_object(augmentation_list)
-        self.data_loader = data_loader
-        self.len = len(self.df)
-        self.image_key = image_key if image_key is not None else 'image'
-        self.mask_key = mask_key if mask_key is not None else 'mask'
-    
-    def __len__(self) -> int:
-        return self.len
-    
-    def load_image(self, idx):
-        image_name = os.path.join(self.root_dir, self.df.iloc[idx][self.image_key])
-        image = Image.open(image_name).convert('RGB')
-        return np.array(image)
 
     def load_polygon(self, idx):
         mask_name = os.path.join(self.root_dir, self.df.iloc[idx][self.mask_key])
@@ -319,13 +350,10 @@ class PolygonRNNDataset(Dataset):
         polygon = np.array(json_file['polygon'])
         point_num = len(json_file['polygon'])
         return polygon, point_num
-
-    def to_tensor(self, x):
-        return x if isinstance(x, torch.Tensor) \
-            else torch.from_numpy(x)
     
     def __getitem__(self, index) -> Dict[str, torch.Tensor]:
-        image = self.load_image(index)
+        image = self.load_image(
+            index, key=self.image_key, is_mask=False, force_rgb=True)
         polygon, num_vertexes = self.load_polygon(index)
         label_array, label_index_array = self.build_arrays(
             polygon, num_vertexes)
@@ -383,6 +411,137 @@ class PolygonRNNDataset(Dataset):
         index = index_b * 28 + index_a
         label_array[point_count, index] = 1
         label_index_array[point_count] = index
+
+class ObjectDetectionDataset(AbstractDataset):
+    def __init__(
+        self,
+        input_csv_path: Path,
+        root_dir=None,
+        augmentation_list=None,
+        data_loader=None,
+        image_key=None,
+        mask_key=None,
+        bounding_box_key=None,
+        n_first_rows_to_read=None,
+        bbox_format='xywh',
+        bbox_output_format='xyxy',
+        bbox_params=None,
+    ) -> None:
+        super(ObjectDetectionDataset, self).__init__(
+            input_csv_path=input_csv_path,
+            root_dir=root_dir,
+            augmentation_list=None,
+            data_loader=data_loader,
+            image_key=image_key,
+            mask_key=mask_key,
+            n_first_rows_to_read=n_first_rows_to_read
+        )
+        self.transform = None if augmentation_list is None \
+            else load_augmentation_object(augmentation_list, bbox_params=bbox_params)
+        self.bounding_box_key = 'bounding_boxes' if bounding_box_key is None else bounding_box_key
+        self.bbox_format = bbox_format
+        self.bbox_output_format = bbox_output_format
+    
+    def convert_bbox(self, bbox):
+        if self.bbox_format == self.bbox_output_format:
+            return bbox
+        elif self.bbox_format == 'xywh' and self.bbox_output_format == 'xyxy':
+            return bbox_xywh_to_xyxy(bbox)
+        elif self.bbox_format == 'xyxy' and self.bbox_output_format == 'xywh':
+            return bbox_xyxy_to_xywh(bbox)
+        else:
+            raise NotImplementedError
+    
+    def load_bounding_boxes_and_labels(self, idx):
+        bbox_path = self.get_path(idx, key=self.bounding_box_key)
+        with open(bbox_path, 'r') as f:
+            json_file = json.load(f)
+        bbox_list, label_list = [], []
+        for box_item in json_file:
+            bbox_list.append(box_item['bbox'])
+            label_list.append(box_item['class'])
+        return torch.as_tensor(bbox_list, dtype=torch.float32), torch.as_tensor(label_list, dtype=torch.int64)
+    
+    def __getitem__(self, index) -> Dict[str, torch.Tensor]:
+        image = self.load_image(
+            index, key=self.image_key, is_mask=False, force_rgb=True)
+        bbox_list, label_list = self.load_bounding_boxes_and_labels(index)
+        ds_item_dict = {
+            'image': image,
+            'bboxes': bbox_list,
+            'labels': label_list,
+        }
+        if self.transform is not None:
+            ds_item_dict = self.transform(**ds_item_dict)
+        image = ds_item_dict.pop('image')
+        ds_item_dict['boxes'] = torch.as_tensor(
+            [
+                self.convert_bbox(bbox) for bbox in ds_item_dict.pop('bboxes')
+            ], dtype=torch.float32)
+        ds_item_dict['labels'] = torch.as_tensor(
+            ds_item_dict['labels'], dtype=torch.int64
+        )
+        return image, ds_item_dict, index
+
+class InstanceSegmentationDataset(ObjectDetectionDataset):
+    def __init__(
+        self,
+        input_csv_path: Path,
+        root_dir=None,
+        augmentation_list=None,
+        data_loader=None,
+        image_key=None,
+        mask_key=None,
+        bounding_box_key=None,
+        n_first_rows_to_read=None,
+        bbox_format='xywh',
+        bbox_output_format='xyxy',
+        return_mask=True,
+        bbox_params=None,
+    ) -> None:
+        mask_key = 'polygon_mask' if mask_key is None else mask_key
+        super(InstanceSegmentationDataset, self).__init__(
+            input_csv_path=input_csv_path,
+            root_dir=root_dir,
+            augmentation_list=augmentation_list,
+            data_loader=data_loader,
+            image_key=image_key,
+            mask_key=mask_key,
+            bounding_box_key=bounding_box_key,
+            n_first_rows_to_read=n_first_rows_to_read,
+            bbox_format=bbox_format,
+            bbox_output_format=bbox_output_format,
+            bbox_params=bbox_params
+        )
+        self.return_mask = return_mask
+
+    def __getitem__(self, index) -> Dict[str, torch.Tensor]:
+        image = self.load_image(
+            index, key=self.image_key, is_mask=False, force_rgb=True)
+        bbox_list, label_list = self.load_bounding_boxes_and_labels(index)
+        ds_item_dict = {
+            'image': image,
+            'bboxes': bbox_list,
+            'labels': label_list,
+        }
+        if self.return_mask:
+            ds_item_dict['masks'] = [self.load_image(
+                index, key=self.mask_key, is_mask=True)]
+        if self.transform is not None:
+            ds_item_dict = self.transform(**ds_item_dict)
+        image = ds_item_dict.pop('image')
+        ds_item_dict['boxes'] = torch.as_tensor(
+            [
+                self.convert_bbox(bbox) for bbox in ds_item_dict.pop('bboxes')
+            ], dtype=torch.float32)
+        ds_item_dict['labels'] = torch.as_tensor(
+            ds_item_dict['labels'], dtype=torch.int64
+        )
+        if self.return_mask:
+            ds_item_dict['masks'] = torch.as_tensor(
+                ds_item_dict['masks'], dtype=torch.uint8
+            )
+        return image, ds_item_dict, index
 
 if __name__ == '__main__':
     pass