+++ tests/test_dataset.py
@@ -20,19 +20,20 @@
  ****
 """
 
+import dataclasses
 import json
 import os
 
 import albumentations as A
-import dataclasses
 import hydra
 import numpy as np
-from hydra.experimental import compose, initialize
+from hydra import compose, initialize
 from pytorch_segmentation_models_trainer.config_definitions.coco_dataset_config import (
     AnnotationConfig, CategoryConfig, CocoDatasetConfig, CocoDatasetInfoConfig,
     ImageConfig, LicenseConfig)
 from pytorch_segmentation_models_trainer.dataset_loader.dataset import (
-    PolygonRNNDataset, SegmentationDataset, load_augmentation_object)
+    InstanceSegmentationDataset, ObjectDetectionDataset, PolygonRNNDataset,
+    SegmentationDataset, load_augmentation_object)
 
 from tests.utils import CustomTestCase
 
@@ -41,6 +42,8 @@ frame_field_root_dir = os.path.join(
     current_dir, 'testing_data', 'data', 'frame_field_data')
 polygon_rnn_root_dir = os.path.join(
     current_dir, 'testing_data', 'data', 'polygon_rnn_data')
+detection_root_dir = os.path.join(
+    current_dir, 'testing_data', 'data', 'detection_data')
 
 class Test_TestDataset(CustomTestCase):
 
@@ -50,7 +53,7 @@ class Test_TestDataset(CustomTestCase):
                 config_name="dataset.yaml",
                 overrides=['input_csv_path='+self.csv_ds_file]
             )
-            ds_from_cfg = hydra.utils.instantiate(cfg)
+            ds_from_cfg = hydra.utils.instantiate(cfg, _recursive_=False)
             ds_from_ref = SegmentationDataset(input_csv_path=self.csv_ds_file)
             self.assertEqual(
                 ds_from_cfg.input_csv_path,
@@ -63,7 +66,7 @@ class Test_TestDataset(CustomTestCase):
                 config_name="dataset.yaml",
                 overrides=['input_csv_path='+self.csv_ds_file]
             )
-            ds_from_cfg = hydra.utils.instantiate(cfg)
+            ds_from_cfg = hydra.utils.instantiate(cfg, _recursive_=False)
             ds_from_ref = SegmentationDataset(input_csv_path=self.csv_ds_file)
             assert(
                 np.array_equal(ds_from_cfg[0]['image'], ds_from_ref[0]['image']) and \
@@ -79,7 +82,7 @@ class Test_TestDataset(CustomTestCase):
                     '+root_dir='+self.root_dir
                 ]
             )
-            ds_from_cfg = hydra.utils.instantiate(cfg)
+            ds_from_cfg = hydra.utils.instantiate(cfg, _recursive_=False)
             ds_from_ref = SegmentationDataset(
                 input_csv_path=self.csv_ds_file_without_root,
                 root_dir=self.root_dir
@@ -105,7 +108,7 @@ class Test_TestDataset(CustomTestCase):
                     '+n_first_rows_to_read=2'
                 ]
             )
-            ds_from_cfg = hydra.utils.instantiate(cfg)
+            ds_from_cfg = hydra.utils.instantiate(cfg, _recursive_=False)
             self.assertEqual(len(ds_from_cfg), 2)
     
     def test_create_frame_field_dataset_instance(self):
@@ -118,7 +121,7 @@ class Test_TestDataset(CustomTestCase):
                     'root_dir='+frame_field_root_dir
                 ]
             )
-            frame_field_ds = hydra.utils.instantiate(cfg)
+            frame_field_ds = hydra.utils.instantiate(cfg, _recursive_=False)
         self.assertEqual(len(frame_field_ds), 12)
         self.assertEqual(frame_field_ds[0]['image'].shape,(571, 571, 3))
         self.assertEqual(frame_field_ds[0]['gt_polygons_image'].shape,(571, 571, 3))
@@ -232,3 +235,36 @@ class Test_TestDataset(CustomTestCase):
             self.assertEqual(ds_item['x2'].shape, (58, 787))
             self.assertEqual(ds_item['x3'].shape, (58, 787))
             self.assertEqual(ds_item['ta'].shape, (58,))
+
+    def test_object_detection_dataset(self):
+        csv_path = os.path.join(detection_root_dir, 'geo', 'dsg_dataset.csv')
+        obj_det_ds = ObjectDetectionDataset(
+            input_csv_path=csv_path,
+            root_dir=os.path.dirname(csv_path),
+            augmentation_list=[
+                A.Normalize(),
+                A.pytorch.ToTensorV2()
+            ]
+        )
+        self.assertEqual(len(obj_det_ds), 12)
+        image, target, id = obj_det_ds[0]
+        self.assertEqual(image.shape, (3, 571, 571))
+        self.assertEqual(target['boxes'].shape, (2, 4))
+        self.assertEqual(target['labels'].shape, (2,))
+
+    def test_instance_segmentation_dataset(self):
+        csv_path = os.path.join(detection_root_dir, 'geo', 'dsg_dataset.csv')
+        obj_det_ds = InstanceSegmentationDataset(
+            input_csv_path=csv_path,
+            root_dir=os.path.dirname(csv_path),
+            augmentation_list=[
+                A.Normalize(),
+                A.pytorch.ToTensorV2()
+            ]
+        )
+        self.assertEqual(len(obj_det_ds), 12)
+        image, target, id = obj_det_ds[0]
+        self.assertEqual(image.shape, (3, 571, 571))
+        self.assertEqual(target['boxes'].shape, (2, 4))
+        self.assertEqual(target['labels'].shape, (2,))
+        self.assertEqual(target['masks'].shape, (1, 571, 571))