+++ tests/test_train.py
@@ -24,7 +24,7 @@ import os
 import subprocess
 
 import pytorch_lightning as pl
-from hydra.experimental import compose, initialize
+from hydra import compose, initialize
 from parameterized import parameterized
 from pytorch_segmentation_models_trainer.train import train